+++ R/cube-residuals.R
@@ -2,7 +2,7 @@
 # the internal chisq tests use rowSums and colSums, but we need
 # the more abstract 'margin' (specifically cubeMarginTable) to
 # get the right numbers for multiple response.
-standardizedMRResiduals <- function (cube, types) {
+standardizedMRResiduals <- function (cube) {
     cube_array <- as.array(noTransforms(cube))
     cube_dims <- dim(cube_array)
 
@@ -54,25 +54,12 @@ setMethod('rstandard', 'CrunchCube', function (model) {
              "slice to evaluate.")
     }
 
-    if (any(vapply(dimensions(model), is.selectedArrayDim, logical(1)))) {
-        # TODO: remove the reference to `as_selected` when that becomes default
-        halt("rstandard is not implemented with CrunchCubes that use selected ",
-             "arrays. Selected arrays have been deprecated, please recreate ",
-             "your cube using `as_selected()` around multiple response variables.")
-    }
-
-    # determine which dimensinos are multiple response, and treat those special
-    # TODO: make this specific to multiple response rather than both cat array and MR
-    types <- vapply(dimensions(model), function(dim) dim$references$type, character(1))
-
-    if (any(types == 'subvariable_items')) {
-        return(standardizedMRResiduals(model, types))
+    if (any(grepl("^mr_", getDimTypes(model)))) {
+        return(standardizedMRResiduals(model))
     } else {
         return(chisq.test(as.array(noTransforms(model)))$stdres)
     }
 })
-
-
 
 # broadcast a vector of values to a matrix with dimensions dims. Similar to
 # but not exactly the same as numpy's `broadcast_to` method.