+++ R/cube-subset.R
@@ -79,6 +79,7 @@ setMethod("[", "CrunchCube", function (x, i, j, ..., drop = TRUE) {
     # This is so that other cube methods, like as.array, don't get confused.
     out@dims[] <- mapply(subsetArrayDimension,
                          dim = x@dims,
+                         dim_type = getDimTypes(x), 
                          idx = translated_index,
                          SIMPLIFY = FALSE)
     # Some dimensions need to be dropped because they were dropped from the
@@ -205,32 +206,33 @@ translateCubeIndex <- function(x, subset, drop) {
 #' @keywords internal
 skipMissingCategories <- function(cube, index){
     visible_cats <- evalUseNA(cube@arrays$count, dims = cube@dims, useNA = cube@useNA)
-    mapply(function(visable, sub){
-        if (identical(sub, "mr_select_drop")) {
-            # select the "Selected" element of the selection dimension.
-            return(c(TRUE, FALSE, FALSE))
-        }
-        if (isTRUE(sub)) {
-            out <- rep(TRUE, length(visable))
-        } else {
-            out <- rep(FALSE, length(visable))
-            out[visable][sub] <- rep(TRUE, length(sub))
-        }
-        return(out)}, visable = visible_cats, sub = index, SIMPLIFY = FALSE)
+    mapply(
+        function (visible, sub) {
+            if (identical(sub, "mr_select_drop")) {
+                # select the "Selected" element of the selection dimension.
+                ## TODO: Don't assume "selected" is position 1; consider an is.selected attr/vector
+                return(c(TRUE, FALSE, FALSE))
+            }
+            if (isTRUE(sub)) {
+                out <- rep(TRUE, length(visible))
+            } else {
+                out <- rep(FALSE, length(visible))
+                out[visible][sub] <- rep(TRUE, length(sub))
+            }
+            return(out)
+        }, visible = visible_cats, sub = index, SIMPLIFY = FALSE)
 }
 
-subsetArrayDimension <- function(dim, idx){
-    dim$any.or.none <- dim$any.or.none[idx]
+subsetArrayDimension <- function(dim, idx, dim_type){
     dim$missing <- dim$missing[idx]
     dim$name <- dim$name[idx]
 
     # subset the category or item metadata
-    if (dim$references$type == "categorical") {
+    if (dim_type %in% c("categorical", "ca_categories")) {
         dim$references$categories <- dim$references$categories[idx]
-    } else if (dim$references$type == "subvariable_items") {
+    } else if (dim_type %in% c("mr_items", "ca_items")) {
         dim$references$subreferences <- dim$references$subreferences[idx]
     }
 
     return(dim)
 }
-