+++ tests/testthat/test-add-variable.R
@@ -180,6 +180,13 @@ test_that("POSTNewVariable rejects invalid categories", {
         "Invalid category names: must be unique")
 })
 
+test_that("checkVarDefErrors errors correctly", {
+    test_errs <- lapply(list("a", "b", 29), function(x) try(log(x), silent = TRUE))
+    expect_error(checkVarDefErrors(test_errs), "The following variable definitions errored on upload: 1, 2")
+    test_errs <- lapply(list(29, 23, 24), function(x) try(log(x), silent = TRUE))
+    expect_silent(checkVarDefErrors(test_errs))
+})
+
 with_mock_crunch({
     ds <- loadDataset("test ds")
     test_that("assignment restrictions", {
@@ -198,14 +205,16 @@ with_mock_crunch({
             'https://app.crunch.io/api/datasets/1/variables/',
             '{"values":5,"type":"numeric","name":"newvar","alias":"newvar"}')
     })
-})
 
-    test_that("checkVarDefErrors errors correctly", {
-        test_errs <- lapply(list("a", "b", 29), function(x) try(log(x), silent = TRUE))
-        expect_error(checkVarDefErrors(test_errs), "The following variable definitions errored on upload: 1, 2")
-        test_errs <- lapply(list(29, 23, 24), function(x) try(log(x), silent = TRUE))
-        expect_silent(checkVarDefErrors(test_errs))
+    test_that("Adding a variable with all missing doesn't send 'values'", {
+        expect_POST(ds$newvar <- NA_real_,
+            'https://app.crunch.io/api/datasets/1/variables/',
+            '{"type":"numeric","name":"newvar","alias":"newvar"}')
+        expect_POST(ds$newvar <- rep(NA_real_, 25),
+            'https://app.crunch.io/api/datasets/1/variables/',
+            '{"type":"numeric","name":"newvar","alias":"newvar"}')
     })
+})
 
 with_test_authentication({
     ds <- newDataset(df)