+++ tests/testthat/test-cube-dims.R
@@ -1,19 +1,21 @@
 context("Cube dimensions")
 
-test_that("getDimType returns the expected cube dimension types", {
-    ca_mr <- loadCube("cubes/catarray-x-mr.json")
-    expect_equivalent(getDimType(ca_mr),
+test_that("getDimTypes returns the expected cube dimension types", {
+    ca_mr <- loadCube(test_path("cubes/catarray-x-mr.json"))
+    expect_equivalent(getDimTypes(ca_mr),
         c("ca_items", "ca_categories", "mr_items", "mr_selections"
         ))
-    cat_cat <- loadCube("cubes/cat-x-cat.json")
-    expect_equivalent(getDimType(cat_cat), c("categorical", "categorical"))
-    cat_mr_mr <- loadCube("cubes/cat-x-mr-x-mr.json")
-    expect_equivalent(getDimType(cat_mr_mr),
+    cat_cat <- loadCube(test_path("cubes/cat-x-cat.json"))
+    expect_equivalent(getDimTypes(cat_cat), c("categorical", "categorical"))
+    ca <- loadCube(test_path("cubes/cat-array.json"))
+    expect_equivalent(getDimTypes(ca), c("ca_items", "ca_categories"))
+    cat_mr_mr <- loadCube(test_path("cubes/cat-x-mr-x-mr.json"))
+    expect_equivalent(getDimTypes(cat_mr_mr),
         c("categorical", "mr_items", "mr_selections", "mr_items",
             "mr_selections")
     )
-    cattarray_cat <- loadCube("cubes/catarray-x-cat.json")
-    expect_equivalent(getDimType(cattarray_cat),
+    cattarray_cat <- loadCube(test_path("cubes/catarray-x-cat.json"))
+    expect_equivalent(getDimTypes(cattarray_cat),
         c("ca_items", "ca_categories", "categorical")
     )
 })
@@ -22,7 +24,7 @@ with_mock_crunch({
     ## Load a ton of cube fixtures via the tab book feature
     ds <- loadDataset("test ds")
     m <- multitables(ds)[[1]]
-    with_POST("https://app.crunch.io/api/datasets/1/multitables/apidocs-as_selected-tabbook/", {
+    with_POST("https://app.crunch.io/api/datasets/1/multitables/apidocs-tabbook/", {
         book <- tabBook(m, data=ds)
     })
     cube <- book[[2]][[2]]
@@ -76,9 +78,26 @@ with_mock_crunch({
                 ""))
     })
 
-    test_that("subvariable/subrefrence methods", {
+    test_that("Variable metadata retrieved from tuples works on the tuples ", {
+        expect_identical(name(ds$gender), "Gender")
+        expect_identical(description(ds$starttime), "Interview Start Time")
+        expect_identical(alias(ds$gender), "gender")
+        expect_identical(id(ds$gender), "66ae9881e3524f7db84970d556c34552")
+        expect_identical(notes(ds$gender), "")
+        expect_identical(notes(ds$birthyr), "Asked instead of age")
+        expect_false(uniformBasis(ds$mymrset))
+    })
+    
+    test_that("Variable metadata from cubes works on the tuples", {
         catarray_x_mr <- loadCube(test_path("cubes/catarray-x-mr.json"))
         cat_array_var <- variables(catarray_x_mr)[[1]]
+        mr_var <- variables(catarray_x_mr)[[3]]
+        
+        expect_identical(name(cat_array_var), "feeling CA")
+        expect_identical(alias(cat_array_var), "feeling_ca")
+        expect_identical(description(cat_array_var), "")
+        expect_identical(notes(cat_array_var), "")
+
         expect_equal(aliases(subvariables(cat_array_var)),
             c("cat_feeling", "dog_feeling"))
         expect_equal(names(subvariables(cat_array_var)),